requirements.txt
asgiref==3.8.1
Django==4.2.16
pillow==11.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
asgiref==3.8.1
Django==4.2.16
django-ckeditor==6.7.1
django-js-asset==2.2.0
pillow==11.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
